chore: move to web3 store
