+++ src/app/reducers.ts
@@ -48,7 +48,6 @@ interface State {
   showNotificationSpotlight: boolean
   hideBalances: boolean
   pendingDeepLinks: PendingDeepLink[]
-  demoModeEnabled: boolean
 }
 
 interface PendingDeepLink {
@@ -90,7 +89,6 @@ const initialState = {
   showNotificationSpotlight: false,
   hideBalances: false,
   pendingDeepLinks: [],
-  demoModeEnabled: false,
 }
 
 function getPersistedDeepLinks(deepLinks: PendingDeepLink[]) {
@@ -258,11 +256,6 @@ export const appReducer = (
         pendingDeepLinks: state.pendingDeepLinks.filter(
           (pendingDeepLink) => pendingDeepLink.url !== action.deepLink
         ),
-      }
-    case Actions.DEMO_MODE_TOGGLED:
-      return {
-        ...state,
-        demoModeEnabled: action.enabled,
       }
     default:
       return state
