+++ src/web3/actions.ts
@@ -1,5 +1,6 @@
 export enum Actions {
   SET_ACCOUNT = 'WEB3/SET_ACCOUNT',
+  DEMO_MODE_TOGGLED = 'WEB3/DEMO_MODE_TOGGLED',
 }
 
 export interface SetAccountAction {
@@ -7,11 +8,23 @@ export interface SetAccountAction {
   address: string
 }
 
-export type ActionTypes = SetAccountAction
+interface DemoModeToggled {
+  type: Actions.DEMO_MODE_TOGGLED
+  enabled: boolean
+}
+
+export type ActionTypes = SetAccountAction | DemoModeToggled
 
 export const setAccount = (address: string): SetAccountAction => {
   return {
     type: Actions.SET_ACCOUNT,
     address: address.toLowerCase(),
+  }
+}
+
+export const demoModeToggled = (enabled: boolean): DemoModeToggled => {
+  return {
+    type: Actions.DEMO_MODE_TOGGLED,
+    enabled,
   }
 }
